Feat : Seven Segment Display
In this commit is the seven segment display for the basys-3 board. Since the board only has the capability to display 4 digits, any number greater than 9999 will be displayed as invalid(----) and first led of the board will light up.
Double dabble algorithm is used to convert the binary number set using switch to decimal number. These numbers are then displayed in the 7 segment display.
Interesting thing about the project was displaying the digit itself. The seven segment display contains 4 anodes and 7 cathodes. The cathod determine the led to be lightened up and the anode determine the digit. So, a different anode is set at each clock pulse (slow clock pulse to ensure the digits are displayed as the clock is too fast for human eyes to see with no visible flicker).